First Areas & Models Implemented
Areas Added:
- Games
This is intended to handle all functionality around each levelpack/campaign. This is where the user will be able to see a list of games. It also might be best to put the leaderboard functionality into this area as well depending on how big the code is for this later on.
Models implemented here:
-- CategoryModel
-- GameModel
-- RulesModel

- Runs
Runs area will hold all the functionality for individual runs, creating them, deleting them and editing them.

--RunModel

These models have been added to a new migration, AddInitialGamesAndRunsModels.