Razor Pages Login GET ported to MVC
I wasn't too sure how the ErrorMessage property of the LoginModel was used so for now I've cut it out. Other functionality will probably be changed later on after this is all merged since theres just better ways for doing these things when we are using MVC.

Will probably move some of the functionality here into a services folder later on to keep it clean.