Leaderboard Model & Repository setup
A lot of back-end stuff included here which would be a bit much for me to write about all of it. 4 models created. Runs, Levels, Categories and Characters. These have all been given repository CRUD functions for the database. Everything is basically in place now to begin work on the leaderboards.