ViewModels & Leaderboard React demo
Got rid of models being used in viewmodels and put everything into viewmodels along with using the automapper to map between them.

Small react setup on the leaderboard index. I want to go with an SPA setup for the leaderboard index so I'm going to be trying to use react for it.